requirements.txt
jinja2
extruct
lxml
beautifulsoup4
urllib3
python-dotenv
streamlit
pyyaml
requests
